Implement global logging in common
